@@ -1,4 +1,4 @@
-import { HttpClient, HttpClientModule } from '@angular/common/http';
+import {HttpClient, HttpClientModule} from '@angular/common/http';
 import {Component, Input} from '@angular/core';
 import { marked } from 'marked';
 import  clip  from 'text-clipper';
@@ -9,7 +9,11 @@ import { Router } from '@angular/router';
 @Component({
   selector: 'app-post-preview',
   standalone: true,
-  imports: [HttpClientModule],
+  imports: [
+// TODO: `HttpClientModule` should not be imported into a component directly.
+// Please refactor the code to add `provideHttpClient()` call to the provider list in the
+// application bootstrap logic and remove the `HttpClientModule` import from this component.
+HttpClientModule],
   providers: [FileService],
   templateUrl: './post-preview.component.html',
   styleUrl: './post-preview.component.css',
